Removed the temporary logout button component and its duplicates within the layout. Logout is now a part of user menu.
